Add PS Vita text effects and 2D render layering support
Simplifies the GXM solid-color program, adds layering to the 2D render
manager with new source-audit coverage, and introduces a text effect
pass. Includes plan/spec docs for Vita text effects and DS/3DS bottom
controls platform cooking, plus a launch-vita3k test script.

+++ src/platform/psvita/rendering/PsVitaQueuedQuad.hpp
@@ -14,8 +14,8 @@ namespace helengine::psvita::rendering {
         /// Stores the PS Vita runtime texture used by this quad.
         PsVitaRuntimeTexture* Texture;
 
-        /// Stores the 2D render-order bucket carried by the originating drawable.
-        std::uint8_t RenderOrder;
+        /// Stores the composite camera and local 2D render order carried by the originating drawable.
+        std::uint16_t RenderOrder;
 
         /// Stores the four textured vertices that describe the quad.
         PsVitaTexturedQuadVertex Vertices[4];
