+++ src/platform/psvita/rendering/PsVitaSolidColorVertex.hpp
@@ -5,12 +5,12 @@
 #include <cstdint>
 
 namespace helengine::psvita::rendering {
-    /// Stores one PS Vita solid-color triangle vertex with screen-space position, packed color, and authored render order.
+    /// Stores one PS Vita solid-color triangle vertex with screen-space position, packed color, and composite render order.
     struct PsVitaSolidColorVertex final {
         float PositionX = 0.0f;
         float PositionY = 0.0f;
         std::uint32_t ColorAbgr = 0u;
-        std::uint8_t RenderOrder = 0u;
+        std::uint16_t RenderOrder = 0u;
     };
 }
